+++ spec/factories/configs.rb
@@ -1,0 +1,6 @@
+FactoryGirl.define do
+  factory :config do
+    subject
+    compatibility { Config::DEFAULT_COMPATIBILITY }
+  end
+end