+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # avro-schema-registry
 
+## v0.4.0
+- Use `avro-salsify-fork` v1.9.0.3.
+- Implement full schema compatibility check and transitive options from
+  Confluent Schema Registry API v3.1.0.
+
 ## v0.3.0
 - Use `heroku_rails_deploy` gem.